requirements.txt
streamlit==1.31.1
pandas==2.2.0
joblib==1.3.2
scikit-learn==1.6.1
xgboost==2.0.3
requests==2.31.0
python-dotenv==1.0.1
